Refactoring in preparation for new .gitignore settings
