+++ test/transformer/regression/arrays/DoubleArraysIndexAssignmentTest1.java
@@ -1,4 +1,4 @@
-package testclasses.arrays;
+package transformer.regression.arrays;
 
 import util.ArgVRandom;
