--- testOutput/testclasses/MinimalSymbolicIntTestCase.java
@@ -1,17 +1,0 @@
-/** filtered and transformed by ARG-V */
- package testclasses;
-
-import gov.nasa.jpf.symbc.Debug;
-
-public class MinimalSymbolicIntTestCase {
-	/** ARG-V: suitable */
-	 public static int baseTestCase(int x) {
-		int replaceTarget = Debug.makeSymbolicInteger("x9") * 10;
-		if (x > replaceTarget) {
-			return 1;
-		} else {
-			return 0;
-		}
-	}
-
-}
